Fix some team colors not working anymore, closes #91

+++ src/rv/util/jogl/MaterialUtil.java
@@ -5,8 +5,8 @@ import js.jogl.model.ObjMaterial;
 
 public class MaterialUtil {
     public static void setColor(ObjMaterial material, Color color) {
-        float[] components = new float[] { color.getRed() / 255, color.getGreen() / 255,
-                color.getBlue() / 255, 1 };
+        float[] components = new float[] { (float) color.getRed() / 255,
+                (float) color.getGreen() / 255, (float) color.getBlue() / 255, 1 };
         material.setAmbient(components);
         material.setDiffuse(components);
     }
